# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "sms-app api documentation",
        "description": "sms-app api documentation",
        "contact": {
            "email": "orutu1@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/groups": {
            "get": {
                "tags": [
                    "Show All Groups"
                ],
                "description": "Show all groups",
                "operationId": "AllGroups",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Group"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sorry no group found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Create Group"
                ],
                "description": "Create new group",
                "operationId": "CreateGroup",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Enter group name",
                        "required": true
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Enter group description",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group created successfully"
                    },
                    "400": {
                        "description": "Sorry unable to create group"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/groups/{id}": {
            "get": {
                "tags": [
                    "Show Group"
                ],
                "description": "Show single group",
                "operationId": "ShowGroup",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Enter group id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Group"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Sorry this data do not exist"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Delete Group"
                ],
                "summary": "Delete existing group",
                "description": "Deletes a group and returns no content",
                "operationId": "deleteGroup",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Group Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Sorry this group do not exist"
                    },
                    "400": {
                        "description": "Sorry unable to delete group"
                    }
                }
            },
            "patch": {
                "tags": [
                    "UpdateGroup"
                ],
                "description": "Update existing group",
                "operationId": "UpdateGroup",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Enter group name",
                        "required": true
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Enter group description",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "group updated successfully"
                    },
                    "400": {
                        "description": "Sorry unable to update group"
                    },
                    "404": {
                        "description": "Sorry this data do not exist"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
        "/recharges": {
            "get": {
                "tags": [
                    "Show All Reacharge"
                ],
                "description": "Show all reacharge",
                "operationId": "AllReacharge",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Recharge"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sorry no recharge found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Create Recharge"
                ],
                "description": "Create new Recharge",
                "operationId": "CreateRecharge",
                "parameters": [
                    {
                        "name": "unit_id",
                        "in": "query",
                        "description": "Enter the unit id",
                        "required": true
                    },
                    {
                        "name": "number",
                        "in": "query",
                        "description": "Enter the number of unit",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recharge successfully"
                    },
                    "400": {
                        "description": "Sorry unable to Recharge"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/recharges/{id}": {
            "get": {
                "tags": [
                    "Show Recharge"
                ],
                "description": "Show single Recharge",
                "operationId": "ShowRecharge",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Enter Recharge id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Sorry this data do not exist"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Delete Recharge"
                ],
                "summary": "Delete existing Recharge",
                "description": "Deletes a Recharge and returns no content",
                "operationId": "deleteRecharge",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Recharge Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Sorry this Recharge do not exist"
                    },
                    "400": {
                        "description": "Sorry unable to delete Recharge"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Group": {
                "title": "Group",
                "description": "Group model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name of the group",
                        "type": "string",
                        "example": "All Staff"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the group",
                        "type": "string",
                        "example": "This is for cars and truck description"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "slug": {
                        "title": "Slug",
                        "description": "Slug",
                        "type": "string",
                        "example": "all-staff"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Group"
                }
            },
...
            "Recharge": {
                "title": "Recharge",
                "description": "Recharge model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "title": "User ID",
                        "description": "User id ",
                        "type": "string",
                        "example": "id"
                    },
                    "unit_id": {
                        "title": "Unit Id",
                        "description": "Unit id",
                        "type": "string",
                        "example": "1"
                    },
                    "number": {
                        "title": "Number",
                        "description": "Number of unit",
                        "type": "string",
                        "example": "100"
                    },
                    "amount": {
                        "title": "Amount",
                        "description": "Unit amount",
                        "type": "string",
                        "example": "1000"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Recharge"
                }
            },
...
        }
    },
...
    "security": [
        []
    ]
}